Remove Usage of Custom JSON Gem
An initial pass at the RPC interface relied on a hacked version of
the JSON Gem, allowing it to return fractional numbers as
BigDecimals instead of Floats, which are prone to precision errors.

An issue (or feature request) has been filed. Until then, a
conversion method has been added to change the Floats to BigDecimals
"by hand".